museume-backend-main/requirements.txt
django==5.1.2
pymysql==1.1.1
python-dotenv==1.0.1
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django_filter==24.3
django-nested-admin==4.1.1
stripe==11.5.0
pillow==11.1.0
gunicorn==23.0.0
sendgrid==6.11.0
cryptography==44.0.2